feat: log decryption in TS purge
