+++ yarn-project/builder/src/contract-interface-gen/typescript.ts
@@ -344,7 +344,6 @@ import {
   type FieldLike,
   Fr,
   type FunctionSelectorLike,
-  L1EventPayload,
   loadContractArtifact,
   loadContractArtifactForPublic,
   type NoirCompiledContract,
