+++ yarn-project/stdlib/package.json
@@ -34,7 +34,6 @@
     "./tx": "./dest/tx/index.js",
     "./fees": "./dest/fees/index.js",
     "./note": "./dest/note/index.js",
-    "./event": "./dest/event/index.js",
     "./p2p": "./dest/p2p/index.js",
     "./errors": "./dest/errors/index.js",
     "./stats": "./dest/stats/index.js",
