--- yarn-project/stdlib/src/event/index.ts
@@ -1,3 +1,0 @@
-export * from './event.js';
-export * from './event_metadata.js';
-export * from './l1_event_payload.js';
