+++ yarn-project/stdlib/src/logs/index.ts
@@ -7,5 +7,5 @@ export * from './log_id.js';
 export * from './log_filter.js';
 export * from './extended_public_log.js';
 export * from './extended_contract_class_log.js';
-export * from './l1_payload/index.js';
+export * from './shared_secret_derivation.js';
 export * from './tx_scoped_l2_log.js';
