--- yarn-project/stdlib/src/logs/l1_payload/index.ts
@@ -1,2 +1,0 @@
-export * from './encrypted_log_payload.js';
-export * from './shared_secret_derivation.js';
